requirements.txt
aiofiles==0.4.0
aiohttp==3.6.2
aiosqlite==0.17.0
aiotask-context==0.6.1
async-timeout==3.0.1
asyncio==3.4.3
asynctest==0.13.0
attrs==22.1.0
certifi==2022.9.14
chardet==3.0.4
coverage==5.0.3
entrypoints==0.3
flake8==3.7.9
idna==2.8
jsonschema==3.2.0
logfmt==0.4
mccabe==0.6.1
multidict==4.7.6
mypy==0.761
mypy-extensions==0.4.3
pycodestyle==2.5.0
pyflakes==2.1.1
pyrsistent==0.18.1
PyYAML==5.3
requests==2.22.0
six==1.16.0
tornado==6.0.3
typed-ast==1.4.3
typing_extensions==4.3.0
urllib3==1.25.11
yarl==1.8.1
